requirements.txt
anyio==3.6.2
autopep8==1.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
mysql-connector-python==8.0.33
numpy==1.24.2
pandas==2.0.0
pip==23.0.1
protobuf==3.20.3
pycodestyle==2.10.0
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
setuptools==66.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.9
starlette==0.26.1
toml==0.10.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.21.1
wheel==0.38.4
SQLAlchemy==1.4.22
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.1
Jinja2==3.1.2
Flask==2.2.0
Flask-Cors==3.0.10